rename node module so that render::node isn't shadowed
